+++ cli/tests/fixtures/blast_radius/blog/views.py
@@ -1,0 +1,5 @@
+from .models import Post
+
+
+def recent(request):
+    return Post.objects.filter(author__id=request.user.id).order_by("-author")